requirements_docker.txt
absl-py==1.4.0
alembic==1.11.1
altair==5.0.1
ansi2html==1.8.0
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.6
boltons 
bottle==0.12.25
brotlipy==0.7.0
cachetools==5.3.1
certifi 
cffi 
charset-normalizer 
click==8.1.6
cloudpickle==2.2.1
cmaes==0.10.0
cmake==3.27.0
colorama==0.4.6
colorlog==6.7.0
comm==0.1.3
conda 
conda-content-trust 
conda-libmamba-solver 
conda-package-handling 
conda_package_streaming 
contourpy==1.1.0
cryptography 
cycler==0.11.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
deepecho==0.0.0
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
Faker==19.1.0
fastjsonschema==2.17.1
filelock==3.12.2
Flask==2.2.5
flatbuffers==23.5.26
flwr==1.4.0
fonttools==4.41.0
fsspec==2023.6.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.56.2
h5py==3.9.0
huggingface-hub==0.16.4
idna 
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.24.0
ipython==8.14.0
iterators==0.0.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpatch 
jsonpointer==2.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kt-legacy==1.0.5
liac-arff==2.5.0
libclang==16.0.6
libmambapy 
lit==16.0.6
llvmlite==0.40.1
loguru==0.7.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
minio==7.1.15
mpmath==1.3.0
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.0.74
openml==0.14.0
opt-einsum==3.3.0
optuna==3.2.0
optuna-dashboard==0.10.3
packaging 
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
platformdirs==3.9.1
plotly==5.15.0
pluggy 
prettytable==3.8.0
prince==0.10.8
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat 
pycparser 
Pygments==2.15.1
PyMySQL==1.1.0
pyOpenSSL 
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks 
python-dateutil==2.8.2
pyts==0.13.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
rdt==1.6.0
referencing==0.30.0
regex==2023.6.3
requests 
requests-oauthlib==1.3.1
retrying==1.3.4
rpds-py==0.9.2
rsa==4.9
ruamel.yaml 
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sdmetrics
six 
slicer==0.0.7
SQLAlchemy==2.0.19
stack-data==0.6.2
sympy==1.12
tabnet==0.1.6
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-addons==0.21.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.13.3
toolz 
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm 
traitlets==5.9.0
transformers==4.31.0
transtab==0.0.5
triton==2.0.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3 
wcwidth==0.2.6
webcolors==1.13
Werkzeug==2.2.3
win32-setctime==1.1.0
wincertstore==0.2
wrapt==1.15.0
xmltodict==0.13.0
yarg==0.1.9
zipp==3.16.2
zstandard 
